Artanise Project Description & ReadMe.
Hello,
2 new Files have been added. One is the description of Artanise app. The other is its ReadMe. 
First, Artanise description file is needed for us and other users to understand what the app is, why is it created, how it works, where is it used and developped, releases deadlines, etc.
Second, ReadMe. Not prior but, needed when the app will be coded. Be conscious that it is coming tho, we'll need it to explain to the user how to download and install the app.
There will be later others files published. Indeed, app code, LICENSE AND AGREEMENT file, and gitignore.
Thank you all moth** ***kers